feat: added buttons and selection functionality.
todo: make buttons selectable with keyboard, midi or mouse and trigger action method within button class when any of the above is done.

note: in prior programs the touchscreen functionality has been poor since when I touch the screen it will automatically click wherever was previously selected, so try and find a fix for that (probably click on mouse button up, then hopefully the touch will move the mouse and when you release your hand it will trigger where the touch goes).