feat: drew a piano at the bottom that acn be played with the mouse.
WIP: adding keyboard and midi control but not currently working. If I use skips in the black_key index then it gets it wrong as its looking for black key which is corect, but hte index is out, if you get rid of this it looks for black keys too often but the indexing is correct.